+++ api/src/app.js
@@ -2,7 +2,7 @@ const express = require('express');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 const morgan = require('morgan');
-const routes = require('./routes/index.js');
+const routes = require('./index.js');
 require('dotenv').config();
 const {CORS} = process.env;
 require('./db.js');
@@ -17,6 +17,8 @@ server.use(bodyParser.urlencoded({ extended: true, limit: '50mb' }));
 server.use(bodyParser.json({ limit: '50mb' }));
 server.use(cookieParser());
 server.use(morgan('dev'));
+
+
 server.use((req, res, next) => {
   res.header('Access-Control-Allow-Origin', CORS); // update to match the domain you will make the request from
   res.header('Access-Control-Allow-Credentials', 'true');
