+++ api/src/source/cloudinarySource.js
@@ -7,4 +7,4 @@ cloudinary.config({
 });
 
 
-module.exports = { cloudinary };
+module.exports =  cloudinary ;
